+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Usage:
 ### 1.16-SNAPSHOT
 * Fix #1690: Base images based on ubi9
 * Fix #2390: support for all missing Chart.yaml fields
+* Fix #2444: Add support for Spring Boot application properties placeholders
 
 ### 1.15.0 (2023-11-10)
 * Fix #2138: Support for Spring Boot Native Image